requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.7.1
dnspython==2.1.0
flake8==3.9.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mccabe==0.6.1
multidict==5.1.0
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.11.3
python-dotenv==0.17.0
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
